Periodic_Updates_24.10.16
1. 방 생성 시 게이트 연결이 안되던 버그 수정
2. 오브젝트 풀링으로 사용된 총알이 안나가는 경우가 발생했었는데, 이를 수정
3. 총알이 파괴되는 판정으로 수정. -> 특정 Tag를 제외한 모든 곳에서 파괴에서 특정 Tag에서만 파괴되도록 (전에는 예외 상황을 너무 많이 만들어줘야하는 단점이 존재하고 특정 버그가 있었음)
4. 간단한 미니맵 구현

+++ Assets/Scripts/Controller/Combat/Bullet/MissileBulletController.cs
@@ -115,6 +115,7 @@ public class MissileBulletController : BulletController
         DOTween.To(() => TargetRange, y => TargetRange = y, ShadowRangeTarget, SpreadTime)
             .SetEase(Ease.Linear);
 
+        ThisRb.simulated = true;
         gameObject.SetActive(true);
     }
 
@@ -122,21 +123,24 @@ public class MissileBulletController : BulletController
 
     #region Collision
     
-    private void OnTriggerEnter2D(Collider2D _Collision)
+    private void OnTriggerEnter2D(Collider2D _Col)
     {
-        if (_Collision.gameObject.tag == "Player" || _Collision.gameObject.tag == "PlayerThing")
+        if (!this.gameObject.activeSelf)
         { return; }
 
         // Hit Enemy
-        if (_Collision.tag == "Enemy")
+        if (_Col.tag == "Enemy")
         {
-            if (_Collision.transform.parent.TryGetComponent(out EnemyController EC))
+            if (_Col.transform.parent.TryGetComponent(out EnemyController EC))
             {
                 EC.TakeDamage(BulletState.DamageType, BulletState.BaseDamage);
             }
         }
 
-        DeleteThis();
+        if (DestroyTagList.Contains(_Col.tag))
+        {
+            DeleteThis();
+        }
     }
 
     #endregion
